+++ front/src/views/About.vue
@@ -10,6 +10,11 @@
         </a>
       </li>
       <li>
+        <a href="https://www.liquidsoap.info/doc-1.4.4/" target="_blank">
+          Liquidsoap documentation
+        </a>
+      </li>
+      <li>
         <a href="https://github.com/martinkirch/showergel" target="_blank">
           Sources and issue tracker
         </a>
@@ -17,3 +22,10 @@
     </ul>
   </div>
 </template>
+
+
+<script>
+export default {
+  props: ["parameters"],
+};
+</script>