+++ showergel/__init__.py
@@ -10,9 +10,10 @@ import sys
 import logging
 import logging.config
 import json
-from configparser import ConfigParser
 from functools import wraps
 
+import toml
+import click
 from bottle import Bottle, response, HTTPError, request, static_file, redirect, HTTPResponse
 from bottle.ext.sqlalchemy import Plugin as SQLAlchemyPlugin
 from sqlalchemy import engine_from_config
@@ -24,35 +25,11 @@ from showergel.liquidsoap_connector import Connection
 _log = logging.getLogger(__name__)
 
 
-def force_python_rootlogger(fn):
-    """
-    Server's logger may bypasses our logging.config.fileConfig,
-    especially when an error occurs. So we add this tiny plugin
-    """
-    @wraps(fn)
-    def _force_python_rootlogger(*args, **kwargs):
-        try:
-            actual_response = fn(*args, **kwargs)
-        except HTTPError:
-            raise
-        except Exception as excn:
-            if isinstance(excn, HTTPResponse):
-                # may happen when redirecting: Bottle raises a response
-                return excn
-            else:
-                _log.exception(excn)
-                raise HTTPError(500, "Internal Error", excn) from None
-        return actual_response
-    return _force_python_rootlogger
-
 def send_cors():
     """
     Send CORS headers along all requests.
     This is only enabled in debug mode, because WebPack's live-compiling-server
     is hosting on another port
-
-    FIXME #12 in some cases (like, errors) it's not returned... could be fixed
-    in Bottle 0.13 - see https://github.com/bottlepy/bottle/issues/1125
     """
     response.headers['Access-Control-Allow-Origin'] = '*'
     response.headers['Access-Control-Allow-Methods'] = 'GET, POST, PUT, OPTIONS, HEAD, DELETE'
@@ -72,7 +49,6 @@ class MainBottle(ShowergelBottle):
 
         http://bottlepy.org/docs/dev/tutorial.html#auto-reloading
         """
-        self.install(force_python_rootlogger)
         if not kwargs['reloader'] or environ.get('BOTTLE_CHILD'):
             self.init(demo=kwargs['demo'], debug=kwargs['debug'])
         elif kwargs['reloader']:
@@ -84,11 +60,12 @@ class MainBottle(ShowergelBottle):
         """
         Showergel's initialization function
         """
+        engine = engine_from_config(self.config, prefix="db.sqlalchemy.")
+        plugin = SQLAlchemyPlugin(engine)
+        self.install(plugin)
         for sub_app in sub_apps:
-            sub_app.config = self.config
-
-        engine = engine_from_config(self.config['db'])
-        self.install(SQLAlchemyPlugin(engine))
+            sub_app.install(plugin)
+            sub_app.config.update(self.config)
 
         if demo:
             self.add_hook('after_request', send_cors)
@@ -99,14 +76,6 @@ class MainBottle(ShowergelBottle):
             self.add_hook('after_request', send_cors)
 
         Connection.setup(self.config)
-
-    def install(self, plugin):
-        """
-        Because Bottle.install does not install plug-ins to sub-applications
-        """
-        super().install(plugin)
-        for sub_app in sub_apps:
-            sub_app.install(plugin)
 
     def get_engine(self):
         for p in self.plugins:
@@ -137,29 +106,29 @@ def enable_cors_generic_route():
     return {}
 
 def read_bool_param(param):
-    value = app.config['listen'].get(param)
+    value = app.config.get('listen.'+param)
     if value in ('False', 'false'):
         return False
     else:
         return bool(value)
 
-def serve():
-    if len(sys.argv) < 2:
-        print("Missing argument: path to showergel's .ini", file=sys.stderr)
-        sys.exit(1)
-    config_path = sys.argv[1]
-    logging.config.fileConfig(config_path, disable_existing_loggers=False)
+@click.command()
+@click.version_option()
+@click.argument('config_path', type=click.Path(readable=True, allow_dash=False, exists=True))
+@click.option('--verbose', is_flag=True, help="Sets logging level to DEBUG")
+def serve(config_path, verbose):
+    with open(config_path, 'r') as f:
+        conf = toml.load(f)
+    logging.config.dictConfig(conf['logging'])
+    if verbose:
+        logging.getLogger().setLevel(logging.DEBUG)
 
-    # we don't use Bottle's app.config.load_config because it's eager loading
-    # values: that doesn't go well with interpolation keys in logging config
-    parser = ConfigParser()
-    parser.read(config_path)
-    app.config = parser
+    app.config.load_dict(conf)
 
     try:
-        port = int(app.config['listen']['port'])
+        port = int(app.config['listen.port'])
     except ValueError:
-        port = int(environ[app.config['listen']['port']])
+        port = int(environ[app.config['listen.port']])
 
     demo = read_bool_param('demo')
     debug = read_bool_param('debug')
@@ -169,7 +138,7 @@ def serve():
         server = 'wsgiref'
     app.run(
         server=server,
-        host=app.config['listen']['address'],
+        host=app.config['listen.address'],
         port=port,
         reloader=debug,
         quiet=True,