+++ showergel/showergel_bottle.py
@@ -1,6 +1,9 @@
 import json
+import logging
 
 from bottle import Bottle, response
+
+_log = logging.getLogger(__name__)
 
 class ShowergelBottle(Bottle):
 
@@ -15,11 +18,7 @@ class ShowergelBottle(Bottle):
             response.status = 400
             return json.dumps({"code": 400, "message": "Please send well-formed JSON"})
         else:
+            if res.exception:
+                _log.critical("Caught exception: %r", res.exception)
+                _log.debug("Caught:", exc_info=res.exception)
             return json.dumps({"code": int(res.status_code), "message": res.body})
-
-    def _handle(self, environ):
-        """
-        Workaround for https://github.com/bottlepy/bottle/issues/602
-        """
-        environ["PATH_INFO"] = environ["PATH_INFO"].encode("utf8").decode("latin1")
-        return super()._handle(environ)