+++ tests/__init__.py
@@ -1,4 +1,5 @@
 import logging
+import sys
 from unittest import TestCase
 
 from webtest import TestApp
@@ -11,12 +12,19 @@ from showergel import app
 from showergel.db import Base
 from showergel.liquidsoap_connector import Connection
 
-app.config = {
-    'db': {
-        "sqlalchemy.url": "sqlite:///:memory:",
+APP_CONFIG = {
+    'db.sqlalchemy': {
+        "url": "sqlite:///:memory:",
         # "echo": True,
+    },
+    'interface': {
+        'name': "ShowergelTest",
+    },
+    'metadata_log': {
+        'extra_fields': ["track*"],
     }
 }
+app.config.load_dict(APP_CONFIG)
 app.init()
 
 __engine = app.get_engine()
@@ -25,9 +33,13 @@ DBSession = sessionmaker(bind=__engine)
 
 bottle.debug(mode=True)
 
+logging.basicConfig(
+    format="%(asctime)s %(levelname)-5.5s [%(name)s:%(lineno)s] %(message)s",
+    level=logging.DEBUG,
+)
+
 class ShowergelTestCase(TestCase):
 
     @classmethod
     def setUpClass(cls):
-        logging.getLogger().setLevel(logging.DEBUG)
         cls.app = TestApp(app)